# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Tripline API
  version: 0.1.0
  description: "Self-hosted dependency drift monitor. Authenticate with the session cookie (browser; state-changing requests need header `X-Tripline-CSRF: 1`) or a read-only bearer API token."
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: tripline_session
    token:
      type: http
      scheme: bearer
      description: Read-only API token (tl_…)
  schemas:
    def-0:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - request_id
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
            details: {}
      title: ErrorResponse
paths:
...
  /api/v1/channels:
    get:
      summary: List notification channels
      tags: &a5
        - channels
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - channels
                properties:
                  channels:
                    type: array
                    items:
                      type: object
                      required:
                        - id
                        - name
                        - kind
                        - enabled
                        - events
                        - target
                        - hasSecret
                        - createdAt
                        - updatedAt
                      properties:
                        id:
                          type: string
                          format: uuid
                        name:
                          type: string
                        kind:
                          enum:
                            - webhook
                            - slack
                        enabled:
                          type: boolean
                        events:
                          type: array
                          items:
                            type: string
                        target:
                          type: string
                        hasSecret:
                          type: boolean
                        createdAt:
                          type: string
                          format: date-time
                        updatedAt:
                          type: string
                          format: date-time
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
    post:
      summary: Create a webhook or Slack channel
      tags: *a5
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - kind
                - url
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 120
                kind:
                  enum:
                    - webhook
                    - slack
                url:
                  type: string
                  minLength: 8
                  maxLength: 2048
                  pattern: ^https?://
                  description: Write-only; stored encrypted.
                secret:
                  type: string
                  minLength: 8
                  maxLength: 256
                  description: HMAC-SHA256 signing secret (webhooks). Write-only.
                events:
                  type: array
                  items:
                    enum:
                      - incident.opened
                      - incident.resolved
                      - drift.breaking
                      - drift.warning
                      - drift.info
                  minItems: 1
                  maxItems: 5
                  uniqueItems: true
                enabled:
                  type: boolean
              additionalProperties: false
      responses:
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - channel
                properties:
                  channel:
                    type: object
                    required:
                      - id
                      - name
                      - kind
                      - enabled
                      - events
                      - target
                      - hasSecret
                      - createdAt
                      - updatedAt
                    properties:
                      id:
                        type: string
                        format: uuid
                      name:
                        type: string
                      kind:
                        enum:
                          - webhook
                          - slack
                      enabled:
                        type: boolean
                      events:
                        type: array
                        items:
                          type: string
                      target:
                        type: string
                      hasSecret:
                        type: boolean
                      createdAt:
                        type: string
                        format: date-time
                      updatedAt:
                        type: string
                        format: date-time
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
  /api/v1/channels/{id}:
    patch:
      summary: "Update a channel (`secret: null` removes the signing secret)"
      tags: *a5
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 120
                url:
                  type: string
                  minLength: 8
                  maxLength: 2048
                  pattern: ^https?://
                  description: Write-only; stored encrypted.
                secret:
                  anyOf:
                    - type: string
                      minLength: 8
                      maxLength: 256
                      description: HMAC-SHA256 signing secret (webhooks). Write-only.
                    - type: "null"
                events:
                  type: array
                  items:
                    enum:
                      - incident.opened
                      - incident.resolved
                      - drift.breaking
                      - drift.warning
                      - drift.info
                  minItems: 1
                  maxItems: 5
                  uniqueItems: true
                enabled:
                  type: boolean
              additionalProperties: false
      parameters:
        - schema:
            type: string
            format: uuid
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - channel
                properties:
                  channel:
                    type: object
                    required:
                      - id
                      - name
                      - kind
                      - enabled
                      - events
                      - target
                      - hasSecret
                      - createdAt
                      - updatedAt
                    properties:
                      id:
                        type: string
                        format: uuid
                      name:
                        type: string
                      kind:
                        enum:
                          - webhook
                          - slack
                      enabled:
                        type: boolean
                      events:
                        type: array
                        items:
                          type: string
                      target:
                        type: string
                      hasSecret:
                        type: boolean
                      createdAt:
                        type: string
                        format: date-time
                      updatedAt:
                        type: string
                        format: date-time
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
    delete:
      summary: Delete a channel
      tags: *a5
      parameters:
        - schema:
            type: string
            format: uuid
          in: path
          name: id
          required: true
      responses:
        "204":
          description: Default Response
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
  /api/v1/channels/{id}/test:
    post:
      summary: Send a test notification synchronously
      tags: *a5
      parameters:
        - schema:
            type: string
            format: uuid
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - ok
                properties:
                  ok:
                    type: boolean
                  status:
                    type: integer
                  error:
                    type: string
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
        "429":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-0"
...
servers:
  - url: /
security:
  - session: []
  - token: []